==> code/0 preprocessing.ipynb <==
# Preprocessing the data

import pandas as pd
import numpy as np

outfit_data = pd.read_csv("../dataset/outfit_data.csv")
product_data = pd.read_csv("../dataset/product_data.csv")

print(outfit_data.shape)
print(product_data.shape)

We discard some categories, as they do not belong to clothes and also categories with really low value counts. 

non_interseting_codes = product_data[product_data['des_product_family'].isin(['Fragances', 'Deco Accessories', 'Intimate', 'Deco Textiles', 'Bedding']) | product_data["des_sex"].isin(["Male","Unisex"]) | product_data["des_age"].isin(["Kids"])]['cod_modelo_color']

product_data = product_data[~product_data['cod_modelo_color'].isin(non_interseting_codes)]
outfit_data = outfit_data[~outfit_data['cod_modelo_color'].isin(non_interseting_codes)]

print(outfit_data.shape)
print(product_data.shape)

### Product data

We imputate color value for the ones that do not have an integer code.

product_data['cod_color_code'] = pd.to_numeric(product_data['cod_color_code'], errors='coerce')

mean_values = product_data.groupby('des_agrup_color_eng')['cod_color_code'].mean()

def replace_nan_with_mean(row):
    if np.isnan(row['cod_color_code']):
        return mean_values[row['des_agrup_color_eng']]
    else:
        return row['cod_color_code']

product_data['cod_color_code'] = product_data.apply(replace_nan_with_mean, axis=1)  

product_data['des_filename'] = product_data['des_filename'].apply(lambda x: x.split('/')[-1])
product_data.head(5)

product_data.isnull().sum()

product_data.columns

product_data['des_line'].value_counts()

product_data['des_fabric'].value_counts()

columns = ['cod_modelo_color', 'cod_color_code', 'des_agrup_color_eng', 'des_fabric', 'des_product_category',
           'des_product_type', 'des_filename', 'des_product_family']

product_data= product_data[columns]
columns_to_encode = product_data.columns.difference(['cod_modelo_color', 'des_filename', 'cod_color_code'])
product_data.to_csv("../dataset/product_data_cleaned.csv")

We perform one-hot encoding for the categorical data

encoded_product_data = pd.get_dummies(product_data, columns=columns_to_encode, dtype=int)
encoded_product_data.shape
encoded_product_data.head(5)

### Outfit data

outfit_data.head(7)

outfit_data.isnull().sum()

### Save preprocessed files

encoded_product_data.to_csv("../dataset/product_data_preprocessed.csv", index=False)
outfit_data.to_csv("../dataset/outfit_data_preprocessed.csv", index=False)
