==> code/2 Image_feature_extraction.ipynb <==
import pandas as pd
import numpy as np
import torch 
from torchvision import models, transforms
from PIL import Image
import os
import matplotlib.pyplot as plt
import matplotlib.image as mpimg
import json
import tqdm


vgg16 = models.vgg16(pretrained=True)

vgg16.classifier = vgg16.classifier[:-1]

vgg16.eval()

def load_and_preprocess_image(image_path):
    image = Image.open(image_path).convert('RGB')

    # Resize while maintaining the original aspect ratio
    aspect_ratio = image.width / image.height
    if aspect_ratio > 1:
        target_size = (int(224 * aspect_ratio), 224)
    else:
        target_size = (224, int(224 / aspect_ratio))

    transform = transforms.Compose([
        transforms.Resize(target_size),
        transforms.CenterCrop(224),
        transforms.ToTensor(),
        transforms.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225]),
    ])

    input_image = transform(image).unsqueeze(0)
    return input_image

preprocessed_image = load_and_preprocess_image('images/2019_41005828_05.jpg')
preprocessed_image.shape

images_folder = 'images'

#warnings off pandas
pd.options.mode.chained_assignment = None

# Iterate over the images in the folder

image_features = pd.DataFrame()
count = 0
dict = {}
for filename in tqdm.tqdm(os.listdir(images_folder)):
    if filename.endswith(".jpg") or filename.endswith(".png"):

        count += 1
    
        # Load and preprocess the image
        image_path = os.path.join(images_folder, filename)
        input_image = load_and_preprocess_image(image_path)

        # Make predictions using the VGG16 model
        with torch.no_grad():
            output = vgg16(input_image)


        output = torch.nn.functional.normalize(output, p=2, dim=1)

        tensor_values = output.view(-1).tolist()

        # Create a dictionary containing the image filename and the tensor values
        dict[filename] = tensor_values

        # Append the dictionary to the image_features DataFrame, use concat



        #image_features = pd.concat([image_features, pd.DataFrame({'filename': filename, 'tensor_values': tensor_values})], ignore_index=True)
        


        # Print the top prediction
        #_, predicted_class = torch.max(output, 1)

        # Map the predicted class index to the label using the loaded JSON file
        #predicted_label = class_labels[predicted_class.item()]

        # Print the results
        #print(f"Image: {filename}, Predicted Class: {predicted_label} (Class Index: {predicted_class.item()})")

data = [{'Filename': filename, **{f'Weight_{i+1}': weight for i, weight in enumerate(weights)}}
        for filename, weights in dict.items()]

# Create a DataFrame from the list of dictionaries
df = pd.DataFrame(data)

df.head()

df.shape

import pickle

#save to a picle
df.to_pickle('image_features.pkl')

#save the df to json
df.to_json('image_features.json', orient='records')
