==> code/3 create_input.ipynb <==
import pandas as pd
import pickle 

product_data = pd.read_csv('../dataset/product_data_preprocessed.csv')
product_data.head()

with open('../dataset/compatibilities.pickle', 'rb') as f:
    compatibilities = pickle.load(f)

labels = pd.DataFrame.from_dict(compatibilities.items())
labels.columns = ['cod_modelo_color', 'community_id']
product_data = labels.merge(product_data, on='cod_modelo_color')
print(product_data.shape)
product_data.head()

with open('../../image_features.pkl', 'rb') as f:
    image_features = pickle.load(f)

image_features.rename(columns={'Filename':'des_filename'}, inplace=True)
product_data = product_data.merge(image_features, on = 'des_filename').drop(['des_filename'], axis=1)

product_data.head()

# dict_modelo_index = dict(zip(product_data.cod_modelo_color, product_data.index))
# dict_index_modelo = dict(zip(product_data.index, product_data.cod_modelo_color))

# with open('../dataset/dict_index_modelo.pickle', 'wb') as f:
#     pickle.dump(dict_index_modelo, f, pickle.HIGHEST_PROTOCOL)

# with open('../dataset/dict_modelo_index.pickle', 'wb') as f:
#     pickle.dump(dict_modelo_index, f, pickle.HIGHEST_PROTOCOL)
    
X = product_data.drop(['community_id'], axis=1)
y = product_data['community_id']

X.shape, y.shape

X.to_csv('../dataset/X.csv', index=False)
y.to_csv('../dataset/y.csv', index=False)
